src/backend/requirements.txt
boto3==1.34.0
botocore==1.34.0
python-jose==3.3.0
requests==2.31.0
python-multipart==0.0.6
pytest==7.4.0
pytest-cov==4.1.0
moto==4.2.0
pytest-mock==3.11.1 